refactor(auth): move module augmentation from plugins into a single declaration override
