+++ src/plugins/db.ts
@@ -6,12 +6,6 @@ import postgres from 'postgres';
 import type { Config } from '../../config/config.d.ts';
 import { users } from '../schema/user.js';
 
-declare module 'fastify' {
-  interface FastifyInstance {
-    db: ReturnType<typeof drizzle<{ users: typeof users }>>;
-  }
-}
-
 const db: FastifyPluginCallback<Config> = (fastify, options, done) => {
   const client = postgres(options.db);
   const db = drizzle(client, {
